Fix launch at login helper.
