+++ MonitorControl/AppDelegate.swift
@@ -59,8 +59,6 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     if !prefs.bool(forKey: Utils.PrefKeys.appAlreadyLaunched.rawValue) {
       prefs.set(true, forKey: Utils.PrefKeys.appAlreadyLaunched.rawValue)
 
-      prefs.set(false, forKey: Utils.PrefKeys.startAtLogin.rawValue)
-
       prefs.set(false, forKey: Utils.PrefKeys.showContrast.rawValue)
       prefs.set(false, forKey: Utils.PrefKeys.lowerContrast.rawValue)
     }
